Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,7 +1,7 @@
  #include <stdio.h>
  int main() {
- int num1,num2;
- scanf("%d %d",&num1,&num2);
- printf("Product: %d",num1 * num2);
+ float num1,num2;
+ scanf("%f %f",&num1,&num2);
+ printf("Product: %.2f",num1 * num2);
      return 0;
 }
